+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0-alpha.4] - 2023-06-07
+- add: `useRaycastMeshes` method:
+  ```ts
+    // call to enable raycasting with low poly raycast meshes
+    // this can be done once in your project
+    useRaycastMeshes(true);
+    
+    // then use the raycaster as usual
+    const raycaster = new Raycaster();
+    raycaster.setFromCamera(mouse, camera);
+    const intersects = raycaster.intersectObjects(scene.children, true);
+    
+    // call to disable raycasting with low polwy meshes
+    useRaycastMeshes(false);
+  ```
+
 ## [1.2.0-alpha.3] - 2023-06-06
 - add: automatically load the highest LOD first to show a slightly better quality level as soon as possible
 - fix: improve Texture LOD selection by taking LOD level height into account